+++ app/src/main/java/ru/ok/itmo/tamtam/ForgotPasswordFragment.kt
@@ -1,4 +1,4 @@
-package ru.ok.itmo.example
+package ru.ok.itmo.tamtam
 
 import androidx.fragment.app.Fragment
