+++ app/src/main/java/ru/ok/itmo/tamtam/Helper.kt
@@ -1,4 +1,4 @@
-package ru.ok.itmo.example
+package ru.ok.itmo.tamtam
 
 import android.content.Context
 import android.widget.Button
