+++ app/src/main/java/ru/ok/itmo/tamtam/MainActivity.kt
@@ -1,4 +1,4 @@
-package ru.ok.itmo.example
+package ru.ok.itmo.tamtam
 
 import android.os.Bundle
 import androidx.activity.viewModels
