+++ app/src/main/res/navigation/login_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/fragmentLogin"
-        android:name="ru.ok.itmo.example.login.LoginFragment"
+        android:name="ru.ok.itmo.tamtam.login.LoginFragment"
         android:label="fragment_login"
         tools:layout="@layout/fragment_login">
         <action
@@ -22,7 +22,7 @@
     <include app:graph="@navigation/app_nav_graph" />
     <fragment
         android:id="@+id/forgotPasswordFragment"
-        android:name="ru.ok.itmo.example.ForgotPasswordFragment"
+        android:name="ru.ok.itmo.tamtam.ForgotPasswordFragment"
         android:label="fragment_forgot_password"
         tools:layout="@layout/fragment_forgot_password" />
 </navigation>
