+++ app/src/main/res/navigation/start_nav_graph.xml
@@ -6,7 +6,7 @@
     app:startDestination="@id/splashFragment">
     <fragment
         android:id="@+id/splashFragment"
-        android:name="ru.ok.itmo.example.SplashFragment"
+        android:name="ru.ok.itmo.tamtam.SplashFragment"
         android:label="fragment_splash"
         tools:layout="@layout/fragment_splash">
         <action
